# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Unified Ads Management SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResponse" }
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [draft, scheduled, publishing, live, partial, failed, paused]
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Campaign" }
    post:
      operationId: createCampaign
      tags: [campaigns]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCampaignRequest" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
  /campaigns/{campaignId}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      parameters:
        - in: path
          name: campaignId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CampaignDetail" }
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      parameters:
        - in: path
          name: campaignId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCampaignRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      parameters:
        - in: path
          name: campaignId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
  /campaigns/{campaignId}/duplicate:
    post:
      operationId: duplicateCampaign
      tags: [campaigns]
      parameters:
        - in: path
          name: campaignId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Duplicated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
...
  /ai/generate-copy:
    post:
      operationId: generateAdCopy
      tags: [ai]
      description: AI-style ad copy generator with per-platform formatting
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GenerateCopyRequest" }
      responses:
        "200":
          description: Generated copy variants
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenerateCopyResponse" }
  /ai/best-time-to-post:
    get:
      operationId: getBestTimeToPost
      tags: [ai]
      parameters:
        - in: query
          name: platform
          schema: { type: string }
      responses:
        "200":
          description: Predicted best posting windows
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BestTimeSlot" }
  /ai/budget-suggestions:
    post:
      operationId: getBudgetSuggestions
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BudgetSuggestionRequest" }
      responses:
        "200":
          description: Suggested budget allocation across platforms
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/BudgetSuggestion" }
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok: { type: boolean }
      required: [ok]
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string }
      required: [id, email, name]
    LoginRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/User" }
      required: [token, user]
...
    Creative:
      type: object
      properties:
        type: { type: string, enum: [image, video, text, carousel] }
        headline: { type: string }
        body: { type: string }
        callToAction: { type: string, nullable: true }
        mediaUrl: { type: string, nullable: true }
        mediaId: { type: string, nullable: true }
      required: [type, headline, body]
    Campaign:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        objective: { type: string, enum: [awareness, traffic, engagement, conversions, leads, app_installs] }
        budget: { type: number }
        currency: { type: string }
        status: { type: string, enum: [draft, scheduled, publishing, live, partial, failed, paused] }
        targetPlatforms:
          type: array
          items: { type: string }
        scheduleStart: { type: string, format: date-time, nullable: true }
        scheduleEnd: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        creative: { $ref: "#/components/schemas/Creative" }
      required: [id, title, objective, budget, currency, status, targetPlatforms, createdAt, updatedAt, creative]
    CampaignDetail:
      type: object
      properties:
        campaign: { $ref: "#/components/schemas/Campaign" }
        jobs:
          type: array
          items: { $ref: "#/components/schemas/PublishJob" }
        logs:
          type: array
          items: { $ref: "#/components/schemas/AuditLog" }
      required: [campaign, jobs, logs]
    CreateCampaignRequest:
      type: object
      properties:
        title: { type: string }
        objective: { type: string }
        budget: { type: number }
        currency: { type: string }
        targetPlatforms:
          type: array
          items: { type: string }
        scheduleStart: { type: string, format: date-time, nullable: true }
        scheduleEnd: { type: string, format: date-time, nullable: true }
        creative: { $ref: "#/components/schemas/Creative" }
        publishImmediately: { type: boolean }
      required: [title, objective, budget, currency, targetPlatforms, creative]
    UpdateCampaignRequest:
      type: object
      properties:
        title: { type: string }
        objective: { type: string }
        budget: { type: number }
        currency: { type: string }
        status: { type: string }
        targetPlatforms:
          type: array
          items: { type: string }
        scheduleStart: { type: string, format: date-time, nullable: true }
        scheduleEnd: { type: string, format: date-time, nullable: true }
        creative: { $ref: "#/components/schemas/Creative" }
...
    PublishJob:
      type: object
      properties:
        id: { type: string }
        campaignId: { type: string }
        platformKey: { type: string }
        platformName: { type: string }
        status: { type: string, enum: [pending, processing, success, failed, retrying] }
        attempts: { type: integer }
        externalId: { type: string, nullable: true }
        externalUrl: { type: string, nullable: true }
        error: { type: string, nullable: true }
        response:
          type: object
          additionalProperties: true
          nullable: true
        impressions: { type: integer }
        clicks: { type: integer }
        spend: { type: number }
        startedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, campaignId, platformKey, platformName, status, attempts, impressions, clicks, spend, createdAt]
...
    AuditLog:
      type: object
      properties:
        id: { type: string }
        campaignId: { type: string, nullable: true }
        platformKey: { type: string, nullable: true }
        action: { type: string }
        level: { type: string, enum: [info, warn, error] }
        message: { type: string }
        meta:
          type: object
          additionalProperties: true
          nullable: true
        createdAt: { type: string, format: date-time }
      required: [id, action, level, message, createdAt]
...
    GenerateCopyRequest:
      type: object
      properties:
        product: { type: string }
        audience: { type: string }
        tone: { type: string, enum: [bold, friendly, professional, playful, urgent] }
        platforms:
          type: array
          items: { type: string }
      required: [product, audience, tone, platforms]
    GenerateCopyResponse:
      type: object
      properties:
        variants:
          type: array
          items: { $ref: "#/components/schemas/CopyVariant" }
      required: [variants]
    CopyVariant:
      type: object
      properties:
        platformKey: { type: string }
        platformName: { type: string }
        headline: { type: string }
        body: { type: string }
        callToAction: { type: string }
        characterCount: { type: integer }
        characterLimit: { type: integer }
      required: [platformKey, platformName, headline, body, callToAction, characterCount, characterLimit]
    BestTimeSlot:
      type: object
      properties:
        platformKey: { type: string }
        platformName: { type: string }
        dayOfWeek: { type: string }
        hour: { type: integer }
        score: { type: number }
        rationale: { type: string }
      required: [platformKey, platformName, dayOfWeek, hour, score, rationale]
    BudgetSuggestionRequest:
      type: object
      properties:
        totalBudget: { type: number }
        objective: { type: string }
        platforms:
          type: array
          items: { type: string }
      required: [totalBudget, objective, platforms]
    BudgetSuggestion:
      type: object
      properties:
        platformKey: { type: string }
        platformName: { type: string }
        suggestedBudget: { type: number }
        sharePct: { type: number }
        rationale: { type: string }
      required: [platformKey, platformName, suggestedBudget, sharePct, rationale]